@@ -52,7 +52,7 @@ public:
   // Read and return the content of a file specified using `embed`.
 };
 
-class Compiler: private SchemaLoader::LazyLoadCallback {
+class Compiler final: private SchemaLoader::LazyLoadCallback {
   // Cross-links separate modules (schema files) and translates them into schema nodes.
   //
   // This class is thread-safe, hence all its methods are const.
