@@ -26,6 +26,7 @@
 
 #ifndef KJ_ASYNC_H_
 #error "Do not include this directly; include kj/async.h."
+#include "async.h"  // help IDE parse this file
 #endif
 
 #ifndef KJ_ASYNC_INL_H_
@@ -254,18 +255,40 @@ class PtmfHelper {
   friend struct GetFunctorStartAddress;
 
 #if __GNUG__
+
   void* ptr;
   ptrdiff_t adj;
   // Layout of a pointer-to-member-function used by GCC and compatible compilers.
+
+  void* apply(void* obj) {
+#if defined(__arm__) || defined(__mips__) || defined(__aarch64__)
+    if (adj & 1) {
+      ptrdiff_t voff = (ptrdiff_t)ptr;
 #else
-#error "TODO(port): PTMF instruction address extraction"
+    ptrdiff_t voff = (ptrdiff_t)ptr;
+    if (voff & 1) {
+      voff &= ~1;
 #endif
+      return *(void**)(*(char**)obj + voff);
+    } else {
+      return ptr;
+    }
+  }
 
 #define BODY \
     PtmfHelper result; \
     static_assert(sizeof(p) == sizeof(result), "unknown ptmf layout"); \
     memcpy(&result, &p, sizeof(result)); \
     return result
+
+#else  // __GNUG__
+
+  void* apply(void* obj) { return nullptr; }
+  // TODO(port):  PTMF instruction address extraction
+
+#define BODY return PtmfHelper{}
+
+#endif  // __GNUG__, else
 
   template <typename R, typename C, typename... P, typename F>
   static PtmfHelper from(F p) { BODY; }
@@ -283,21 +306,6 @@ class PtmfHelper {
   // guess at P. Luckily, if the function parameters are template parameters then it's not
   // necessary to be precise about P.
 #undef BODY
-
-  void* apply(void* obj) {
-#if defined(__arm__) || defined(__mips__) || defined(__aarch64__)
-    if (adj & 1) {
-      ptrdiff_t voff = (ptrdiff_t)ptr;
-#else
-    ptrdiff_t voff = (ptrdiff_t)ptr;
-    if (voff & 1) {
-      voff &= ~1;
-#endif
-      return *(void**)(*(char**)obj + voff);
-    } else {
-      return ptr;
-    }
-  }
 };
 
 template <typename... ParamTypes>
@@ -444,6 +452,28 @@ public:
   }
 };
 
+template <typename T, size_t index>
+class SplitBranch final: public ForkBranchBase {
+  // A PromiseNode that implements one branch of a fork -- i.e. one of the branches that receives
+  // a const reference.
+
+public:
+  SplitBranch(Own<ForkHubBase>&& hub): ForkBranchBase(kj::mv(hub)) {}
+
+  typedef kj::Decay<decltype(kj::get<index>(kj::instance<T>()))> Element;
+
+  void get(ExceptionOrValue& output) noexcept override {
+    ExceptionOr<T>& hubResult = getHubResultRef().template as<T>();
+    KJ_IF_MAYBE(value, hubResult.value) {
+      output.as<Element>().value = kj::mv(kj::get<index>(*value));
+    } else {
+      output.as<Element>().value = nullptr;
+    }
+    output.exception = hubResult.exception;
+    releaseHub(output);
+  }
+};
+
 // -------------------------------------------------------------------
 
 class ForkHubBase: public Refcounted, protected Event {
@@ -479,8 +509,24 @@ public:
     return Promise<_::UnfixVoid<T>>(false, kj::heap<ForkBranch<T>>(addRef(*this)));
   }
 
+  _::SplitTuplePromise<T> split() {
+    return splitImpl(MakeIndexes<tupleSize<T>()>());
+  }
+
 private:
   ExceptionOr<T> result;
+
+  template <size_t... indexes>
+  _::SplitTuplePromise<T> splitImpl(Indexes<indexes...>) {
+    return kj::tuple(addSplit<indexes>()...);
+  }
+
+  template <size_t index>
+  Promise<JoinPromises<typename SplitBranch<T, index>::Element>> addSplit() {
+    return Promise<JoinPromises<typename SplitBranch<T, index>::Element>>(
+        false, maybeChain(kj::heap<SplitBranch<T, index>>(addRef(*this)),
+                          implicitCast<typename SplitBranch<T, index>::Element*>(nullptr)));
+  }
 };
 
 inline ExceptionOrValue& ForkBranchBase::getHubResultRef() {
@@ -585,7 +631,7 @@ private:
   uint countLeft;
   OnReadyEvent onReadyEvent;
 
-  class Branch: public Event {
+  class Branch final: public Event {
   public:
     Branch(ArrayJoinPromiseNodeBase& joinNode, Own<PromiseNode> dependency,
            ExceptionOrValue& output);
@@ -823,6 +869,26 @@ T Promise<T>::wait(WaitScope& waitScope) {
   }
 }
 
+template <>
+inline void Promise<void>::wait(WaitScope& waitScope) {
+  // Override <void> case to use throwRecoverableException().
+
+  _::ExceptionOr<_::Void> result;
+
+  waitImpl(kj::mv(node), result, waitScope);
+
+  if (result.value != nullptr) {
+    KJ_IF_MAYBE(exception, result.exception) {
+      throwRecoverableException(kj::mv(*exception));
+    }
+  } else KJ_IF_MAYBE(exception, result.exception) {
+    throwRecoverableException(kj::mv(*exception));
+  } else {
+    // Result contained neither a value nor an exception?
+    KJ_UNREACHABLE;
+  }
+}
+
 template <typename T>
 ForkedPromise<T> Promise<T>::fork() {
   return ForkedPromise<T>(false, refcounted<_::ForkHub<_::FixVoid<T>>>(kj::mv(node)));
@@ -831,6 +897,11 @@ ForkedPromise<T> Promise<T>::fork() {
 template <typename T>
 Promise<T> ForkedPromise<T>::addBranch() {
   return hub->addBranch();
+}
+
+template <typename T>
+_::SplitTuplePromise<T> Promise<T>::split() {
+  return refcounted<_::ForkHub<_::FixVoid<T>>>(kj::mv(node))->split();
 }
 
 template <typename T>
