@@ -50,16 +50,13 @@ public:
 
   void detach();
   // Don't join the thread in ~Thread().
-  //
-  // TODO(soon): Currently broken: the thread uses the Thread objects during its execution; instead
-  //   the Thread object and the thread itself will need to share a refcounted object.
 
 private:
   struct ThreadState {
     Function<void()> func;
     kj::Maybe<kj::Exception> exception;
 
-    int refcount;
+    unsigned int refcount;
     // Owned by the parent thread and the child thread.
 
     void unref();
